* Lots of 64-bit fixes here and there (invalid int/pointer casts, invalid     mixes of LONG/long/uint32_t, invalid mixes of SHORT/short/uint16_t).   * Fix signed/unsigned compiler warnings.

+++ etw/gfx.c
@@ -598,9 +598,9 @@ AnimObj *LoadAnimObject(char *name, LONG * pens)
 			}
 
 			if ((obj->Frames = calloc(obj->num_frames, sizeof(APTR)))) {
-				if ((obj->Widths = malloc(obj->num_frames * sizeof(LONG)))) {
+				if ((obj->Widths = malloc(obj->num_frames * sizeof(int)))) {
 					if ((obj->Heights =
-						malloc(obj->num_frames * sizeof(LONG)))) {
+						malloc(obj->num_frames * sizeof(int)))) {
 						BOOL ok = TRUE;
 						unsigned short tempw;
